refactor: 소켓 hook 분리
